Espace Vendeur - Dashboard

+++ backend/apps/vendors/serializers.py
@@ -51,7 +51,7 @@ class VendorStatsSerializer(serializers.Serializer):
     total_products = serializers.IntegerField()
     active_products = serializers.IntegerField()
     total_orders = serializers.IntegerField()
-    total_revenue = serializers.DecimalField(max_digits=10, decimal_places=2)
+    total_revenue = serializers.FloatField()
 
 
 #  SERIALIZERS COMMANDES
